Reduce backend memory: single-slot per-hour cache (not 24), discard scored DataFrame after attaching weights, float32 edge-scoring arrays, explicit gc.collect() after startup and scoring

+++ backend/app/shadows.py
@@ -151,7 +151,7 @@ def prepare_shade_inputs(
 
     return ShadeInputs(
         bld_geoms=np.array(bgeoms, dtype=object),
-        bld_heights=np.array(bheights, dtype=float),
+        bld_heights=np.array(bheights, dtype=np.float32),
         tree_canopies=np.array(tcanopies, dtype=object),
     )
